Move all Assignment Progress files into own folder
